requirements.txt
Django==4.2.16
django-debug-toolbar==4.4.6
django-filter==24.3
django-phonenumber-field==8.0.0
django-versatileimagefield==3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Markdown==3.7

